third_party/WebKit/LayoutTests/external/wpt/html/semantics/embedded-content/image-maps/image-map-processing-model/hash-name-reference-expected.txt

This is a testharness.js-based test.
Found 168 tests; 89 PASS, 79 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS HTML (standards) IMG usemap="no-hash-name"
PASS HTML (standards) OBJECT usemap="no-hash-name"
PASS HTML (standards) IMG usemap="no-hash-id"
PASS HTML (standards) OBJECT usemap="no-hash-id"
PASS HTML (standards) IMG usemap="#hash-name"
FAIL HTML (standards) OBJECT usemap="#hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-na... but got Element node <object data="/images/threecolors.png" usemap="#hash-name...
FAIL HTML (standards) IMG usemap="#hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-id... but got Element node <img src="/images/threecolors.png" usemap="#hash-id"></img>
FAIL HTML (standards) OBJECT usemap="#hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-id... but got Element node <object data="/images/threecolors.png" usemap="#hash-id">...
PASS HTML (standards) IMG usemap="#non-map-with-this-name"
FAIL HTML (standards) OBJECT usemap="#non-map-with-this-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <object data="/images/threecolors.png" usemap="#non-map-w...
FAIL HTML (standards) IMG usemap="#non-map-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <img src="/images/threecolors.png" usemap="#non-map-with-...
FAIL HTML (standards) OBJECT usemap="#non-map-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <object data="/images/threecolors.png" usemap="#non-map-w...
PASS HTML (standards) IMG usemap="#two-maps-with-this-name"
FAIL HTML (standards) OBJECT usemap="#two-maps-with-this-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
FAIL HTML (standards) IMG usemap="#two-maps-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <img src="/images/threecolors.png" usemap="#two-maps-with...
FAIL HTML (standards) OBJECT usemap="#two-maps-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
PASS HTML (standards) IMG usemap="#two-maps-with-this-name-or-id"
FAIL HTML (standards) OBJECT usemap="#two-maps-with-this-name-or-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
FAIL HTML (standards) IMG usemap="#two-maps-with-this-id-or-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map...
FAIL HTML (standards) OBJECT usemap="#two-maps-with-this-id-or-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
FAIL HTML (standards) IMG usemap="hash-last#" assert_equals: expected Element node <img src="/images/threecolors.png" usemap="hash-last#"></... but got Element node <area shape="rect" coords="0,0,99,50" href="#area-empty-u...
PASS HTML (standards) OBJECT usemap="hash-last#"
PASS HTML (standards) IMG usemap=""
PASS HTML (standards) OBJECT usemap=""
FAIL HTML (standards) IMG usemap="#" assert_equals: expected Element node <img src="/images/threecolors.png" usemap="#"></img> but got Element node <area shape="rect" coords="0,0,99,50" href="#area-empty-u...
PASS HTML (standards) OBJECT usemap="#"
PASS HTML (standards) IMG usemap="# "
FAIL HTML (standards) OBJECT usemap="# " assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# "></object>
FAIL HTML (standards) IMG usemap="#\n" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-LF... but got Element node <img src="/images/threecolors.png" usemap="#
"></img>
FAIL HTML (standards) OBJECT usemap="#\n" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-LF... but got Element node <object data="/images/threecolors.png" usemap="#
"></object>
PASS HTML (standards) IMG usemap="#percent-escape-name-%41"
PASS HTML (standards) OBJECT usemap="#percent-escape-name-%41"
PASS HTML (standards) IMG usemap="#percent-escape-id-%41"
PASS HTML (standards) OBJECT usemap="#percent-escape-id-%41"
PASS HTML (standards) IMG usemap="#percent-escape-name-%42"
FAIL HTML (standards) OBJECT usemap="#percent-escape-name-%42" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-percent... but got Element node <object data="/images/threecolors.png" usemap="#percent-e...
FAIL HTML (standards) IMG usemap="#percent-escape-id-%42" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-percent... but got Element node <img src="/images/threecolors.png" usemap="#percent-escap...
FAIL HTML (standards) OBJECT usemap="#percent-escape-id-%42" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-percent... but got Element node <object data="/images/threecolors.png" usemap="#percent-e...
PASS HTML (standards) IMG usemap="# hash-space-name"
FAIL HTML (standards) OBJECT usemap="# hash-space-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# hash-spa...
FAIL HTML (standards) IMG usemap="# hash-space-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <img src="/images/threecolors.png" usemap="# hash-space-i...
FAIL HTML (standards) OBJECT usemap="# hash-space-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# hash-spa...
PASS HTML (standards) IMG usemap=" #space-before-hash-name"
FAIL HTML (standards) OBJECT usemap=" #space-before-hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-space-b... but got Element node <object data="/images/threecolors.png" usemap=" #space-be...
FAIL HTML (standards) IMG usemap=" #space-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-space-b... but got Element node <img src="/images/threecolors.png" usemap=" #space-before...
FAIL HTML (standards) OBJECT usemap=" #space-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-space-b... but got Element node <object data="/images/threecolors.png" usemap=" #space-be...
PASS HTML (standards) IMG usemap="http://example.org/#garbage-before-hash-name"
FAIL HTML (standards) OBJECT usemap="http://example.org/#garbage-before-hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-garbage... but got Element node <object data="/images/threecolors.png" usemap="http://exa...
FAIL HTML (standards) IMG usemap="http://example.org/#garbage-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-garbage... but got Element node <img src="/images/threecolors.png" usemap="http://example...
FAIL HTML (standards) OBJECT usemap="http://example.org/#garbage-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-garbage... but got Element node <object data="/images/threecolors.png" usemap="http://exa...
PASS HTML (standards) IMG usemap="#no-such-map"
PASS HTML (standards) OBJECT usemap="#no-such-map"
PASS HTML (standards) IMG usemap="#different-CASE-name"
PASS HTML (standards) OBJECT usemap="#different-CASE-name"
PASS HTML (standards) IMG usemap="#different-CASE-id"
PASS HTML (standards) OBJECT usemap="#different-CASE-id"
PASS HTML (quirks) IMG usemap="no-hash-name"
PASS HTML (quirks) OBJECT usemap="no-hash-name"
PASS HTML (quirks) IMG usemap="no-hash-id"
PASS HTML (quirks) OBJECT usemap="no-hash-id"
PASS HTML (quirks) IMG usemap="#hash-name"
FAIL HTML (quirks) OBJECT usemap="#hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-na... but got Element node <object data="/images/threecolors.png" usemap="#hash-name...
FAIL HTML (quirks) IMG usemap="#hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-id... but got Element node <img src="/images/threecolors.png" usemap="#hash-id"></img>
FAIL HTML (quirks) OBJECT usemap="#hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-id... but got Element node <object data="/images/threecolors.png" usemap="#hash-id">...
PASS HTML (quirks) IMG usemap="#non-map-with-this-name"
FAIL HTML (quirks) OBJECT usemap="#non-map-with-this-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <object data="/images/threecolors.png" usemap="#non-map-w...
FAIL HTML (quirks) IMG usemap="#non-map-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <img src="/images/threecolors.png" usemap="#non-map-with-...
FAIL HTML (quirks) OBJECT usemap="#non-map-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <object data="/images/threecolors.png" usemap="#non-map-w...
PASS HTML (quirks) IMG usemap="#two-maps-with-this-name"
FAIL HTML (quirks) OBJECT usemap="#two-maps-with-this-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
FAIL HTML (quirks) IMG usemap="#two-maps-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <img src="/images/threecolors.png" usemap="#two-maps-with...
FAIL HTML (quirks) OBJECT usemap="#two-maps-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
PASS HTML (quirks) IMG usemap="#two-maps-with-this-name-or-id"
FAIL HTML (quirks) OBJECT usemap="#two-maps-with-this-name-or-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
FAIL HTML (quirks) IMG usemap="#two-maps-with-this-id-or-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map...
FAIL HTML (quirks) OBJECT usemap="#two-maps-with-this-id-or-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
FAIL HTML (quirks) IMG usemap="hash-last#" assert_equals: expected Element node <img src="/images/threecolors.png" usemap="hash-last#"></... but got Element node <area shape="rect" coords="0,0,99,50" href="#area-empty-u...
PASS HTML (quirks) OBJECT usemap="hash-last#"
PASS HTML (quirks) IMG usemap=""
PASS HTML (quirks) OBJECT usemap=""
FAIL HTML (quirks) IMG usemap="#" assert_equals: expected Element node <img src="/images/threecolors.png" usemap="#"></img> but got Element node <area shape="rect" coords="0,0,99,50" href="#area-empty-u...
PASS HTML (quirks) OBJECT usemap="#"
PASS HTML (quirks) IMG usemap="# "
FAIL HTML (quirks) OBJECT usemap="# " assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# "></object>
FAIL HTML (quirks) IMG usemap="#\n" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-LF... but got Element node <img src="/images/threecolors.png" usemap="#
"></img>
FAIL HTML (quirks) OBJECT usemap="#\n" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-LF... but got Element node <object data="/images/threecolors.png" usemap="#
"></object>
PASS HTML (quirks) IMG usemap="#percent-escape-name-%41"
PASS HTML (quirks) OBJECT usemap="#percent-escape-name-%41"
PASS HTML (quirks) IMG usemap="#percent-escape-id-%41"
PASS HTML (quirks) OBJECT usemap="#percent-escape-id-%41"
PASS HTML (quirks) IMG usemap="#percent-escape-name-%42"
FAIL HTML (quirks) OBJECT usemap="#percent-escape-name-%42" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-percent... but got Element node <object data="/images/threecolors.png" usemap="#percent-e...
FAIL HTML (quirks) IMG usemap="#percent-escape-id-%42" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-percent... but got Element node <img src="/images/threecolors.png" usemap="#percent-escap...
FAIL HTML (quirks) OBJECT usemap="#percent-escape-id-%42" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-percent... but got Element node <object data="/images/threecolors.png" usemap="#percent-e...
PASS HTML (quirks) IMG usemap="# hash-space-name"
FAIL HTML (quirks) OBJECT usemap="# hash-space-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# hash-spa...
FAIL HTML (quirks) IMG usemap="# hash-space-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <img src="/images/threecolors.png" usemap="# hash-space-i...
FAIL HTML (quirks) OBJECT usemap="# hash-space-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# hash-spa...
PASS HTML (quirks) IMG usemap=" #space-before-hash-name"
FAIL HTML (quirks) OBJECT usemap=" #space-before-hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-space-b... but got Element node <object data="/images/threecolors.png" usemap=" #space-be...
FAIL HTML (quirks) IMG usemap=" #space-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-space-b... but got Element node <img src="/images/threecolors.png" usemap=" #space-before...
FAIL HTML (quirks) OBJECT usemap=" #space-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-space-b... but got Element node <object data="/images/threecolors.png" usemap=" #space-be...
PASS HTML (quirks) IMG usemap="http://example.org/#garbage-before-hash-name"
FAIL HTML (quirks) OBJECT usemap="http://example.org/#garbage-before-hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-garbage... but got Element node <object data="/images/threecolors.png" usemap="http://exa...
FAIL HTML (quirks) IMG usemap="http://example.org/#garbage-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-garbage... but got Element node <img src="/images/threecolors.png" usemap="http://example...
FAIL HTML (quirks) OBJECT usemap="http://example.org/#garbage-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-garbage... but got Element node <object data="/images/threecolors.png" usemap="http://exa...
PASS HTML (quirks) IMG usemap="#no-such-map"
PASS HTML (quirks) OBJECT usemap="#no-such-map"
PASS HTML (quirks) IMG usemap="#different-CASE-name"
PASS HTML (quirks) OBJECT usemap="#different-CASE-name"
PASS HTML (quirks) IMG usemap="#different-CASE-id"
PASS HTML (quirks) OBJECT usemap="#different-CASE-id"
PASS XHTML img usemap="no-hash-name"
PASS XHTML object usemap="no-hash-name"
PASS XHTML img usemap="no-hash-id"
PASS XHTML object usemap="no-hash-id"
PASS XHTML img usemap="#hash-name"
FAIL XHTML object usemap="#hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-na... but got Element node <object data="/images/threecolors.png" usemap="#hash-name...
PASS XHTML img usemap="#hash-id"
FAIL XHTML object usemap="#hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-id... but got Element node <object data="/images/threecolors.png" usemap="#hash-id">...
PASS XHTML img usemap="#non-map-with-this-name"
FAIL XHTML object usemap="#non-map-with-this-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <object data="/images/threecolors.png" usemap="#non-map-w...
FAIL XHTML img usemap="#non-map-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <img src="/images/threecolors.png" usemap="#non-map-with-...
FAIL XHTML object usemap="#non-map-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-non-map... but got Element node <object data="/images/threecolors.png" usemap="#non-map-w...
PASS XHTML img usemap="#two-maps-with-this-name"
FAIL XHTML object usemap="#two-maps-with-this-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
PASS XHTML img usemap="#two-maps-with-this-id"
FAIL XHTML object usemap="#two-maps-with-this-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
PASS XHTML img usemap="#two-maps-with-this-name-or-id"
FAIL XHTML object usemap="#two-maps-with-this-name-or-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
PASS XHTML img usemap="#two-maps-with-this-id-or-name"
FAIL XHTML object usemap="#two-maps-with-this-id-or-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-two-map... but got Element node <object data="/images/threecolors.png" usemap="#two-maps-...
FAIL XHTML img usemap="hash-last#" assert_equals: expected Element node <img src="/images/threecolors.png" usemap="hash-last#"></... but got Element node <area shape="rect" coords="0,0,99,50" href="#area-empty-u...
PASS XHTML object usemap="hash-last#"
PASS XHTML img usemap=""
PASS XHTML object usemap=""
FAIL XHTML img usemap="#" assert_equals: expected Element node <img src="/images/threecolors.png" usemap="#"></img> but got Element node <area shape="rect" coords="0,0,99,50" href="#area-empty-u...
PASS XHTML object usemap="#"
PASS XHTML img usemap="# "
FAIL XHTML object usemap="# " assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# "></object>
PASS XHTML img usemap="#\n"
FAIL XHTML object usemap="#\n" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-LF... but got Element node <object data="/images/threecolors.png" usemap="#
"></object>
PASS XHTML img usemap="#percent-escape-name-%41"
PASS XHTML object usemap="#percent-escape-name-%41"
PASS XHTML img usemap="#percent-escape-id-%41"
PASS XHTML object usemap="#percent-escape-id-%41"
PASS XHTML img usemap="#percent-escape-name-%42"
FAIL XHTML object usemap="#percent-escape-name-%42" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-percent... but got Element node <object data="/images/threecolors.png" usemap="#percent-e...
PASS XHTML img usemap="#percent-escape-id-%42"
FAIL XHTML object usemap="#percent-escape-id-%42" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-percent... but got Element node <object data="/images/threecolors.png" usemap="#percent-e...
PASS XHTML img usemap="# hash-space-name"
FAIL XHTML object usemap="# hash-space-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# hash-spa...
PASS XHTML img usemap="# hash-space-id"
FAIL XHTML object usemap="# hash-space-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-hash-sp... but got Element node <object data="/images/threecolors.png" usemap="# hash-spa...
PASS XHTML img usemap=" #space-before-hash-name"
FAIL XHTML object usemap=" #space-before-hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-space-b... but got Element node <object data="/images/threecolors.png" usemap=" #space-be...
PASS XHTML img usemap=" #space-before-hash-id"
FAIL XHTML object usemap=" #space-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-space-b... but got Element node <object data="/images/threecolors.png" usemap=" #space-be...
PASS XHTML img usemap="http://example.org/#garbage-before-hash-name"
FAIL XHTML object usemap="http://example.org/#garbage-before-hash-name" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-garbage... but got Element node <object data="/images/threecolors.png" usemap="http://exa...
PASS XHTML img usemap="http://example.org/#garbage-before-hash-id"
FAIL XHTML object usemap="http://example.org/#garbage-before-hash-id" assert_equals: expected Element node <area shape="rect" coords="0,0,99,50" href="#area-garbage... but got Element node <object data="/images/threecolors.png" usemap="http://exa...
PASS XHTML img usemap="#no-such-map"
PASS XHTML object usemap="#no-such-map"
PASS XHTML img usemap="#different-CASE-name"
PASS XHTML object usemap="#different-CASE-name"
PASS XHTML img usemap="#different-CASE-id"
PASS XHTML object usemap="#different-CASE-id"
Harness: the test ran to completion.
